# modules/authn/internal/api/handlers.go
package api

import (
	"github.com/tentackles/shield/modules/authn/internal/auth"
	// "github.com/tentackles/shield/modules/authn/internal/organization" // Placeholder for OrgService
	// "github.com/tentackles/shield/pkg/errors" // Placeholder for ErrorHandler
)

// AuthHandler holds dependencies for authentication and authorization handlers.
type AuthHandler struct {
	authService *auth.AuthService
	// orgService  *organization.Service
	// errorHandler *errors.Handler
	// Add other necessary services like NonceValidator, SessionManager etc.
}

// NewAuthHandler creates and returns a new AuthHandler.
func NewAuthHandler(authService *auth.AuthService /*, orgService *organization.Service, errorHandler *errors.Handler*/) *AuthHandler {
	return &AuthHandler{
		authService: authService,
		// orgService:  orgService,
		// errorHandler: errorHandler,
	}
}

// Placeholder for ErrorResponse and SuccessResponse if they are not globally defined
// type ErrorResponse struct {
// 	Error string `json:"error"`
// }

// type SuccessResponse struct {
// 	Message string `json:"message"`
// }